pyproject.toml
[build-system]
requires = ["setuptools>=75.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "janus"
version = "0.1.0"
description = "Metadata-driven Spark ingestion framework for Brazilian federal open data."
readme = "README.md"
requires-python = ">=3.13,<3.15"
authors = [{ name = "JANUS contributors" }]
dependencies = [
  "PyYAML==6.0.3",
  "pyspark==4.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest==9.0.2",
  "ruff==0.11.6",
]

[project.scripts]
janus = "janus.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-ra"

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["B", "E", "F", "I", "UP"]
